+++ projects/toolchain/jest.config.ci.js
@@ -1,0 +1,12 @@
+// @ts-check
+module.exports = require('./jest.config');
+module.exports.reporters = [
+	'default',
+	[
+		'jest-sonar',
+		{
+			outputDirectory: '../../coverage/cli',
+			outputName: 'sqr.xml'
+		}
+	]
+];